refacotr(WebpackSSR): Move webpack stuff to webpack folder & create client & server separate configs
