+++ src/commands/build.js
@@ -2,8 +2,7 @@ import { resolve } from 'path';
 import promisify from 'es6-promisify';
 import rimraf from 'rimraf';
 import asyncCommand from '../lib/async-command';
-import webpackConfig from '../lib/webpack-config';
-import runWebpack, { showStats, writeJsonStats } from '../lib/run-webpack';
+import runWebpack, { showStats, writeJsonStats } from '../lib/webpack/run-webpack';
 
 export default asyncCommand({
 	command: 'build [src] [dest]',
@@ -39,14 +38,12 @@ export default asyncCommand({
 	},
 
 	async handler(argv) {
-		let config = webpackConfig(argv);
-
 		if (argv.clean) {
 			let dest = resolve(argv.cwd || process.cwd(), argv.dest || 'build');
 			await promisify(rimraf)(dest);
 		}
 
-		let stats = await runWebpack(false, argv, config);
+		let stats = await runWebpack(false, argv);
 		showStats(stats);
 
 		if (argv.json) {
