+++ src/commands/watch.js
@@ -1,6 +1,5 @@
 import asyncCommand from '../lib/async-command';
-import webpackConfig from '../lib/webpack-config';
-import runWebpack, { showStats } from '../lib/run-webpack';
+import runWebpack, { showStats } from '../lib/webpack/run-webpack';
 
 export default asyncCommand({
 	command: 'watch [src]',
@@ -30,9 +29,8 @@ export default asyncCommand({
 
 	async handler(argv) {
 		argv.production = false;
-		let config = webpackConfig(argv);
 
-		let stats = await runWebpack(true, config, showStats);
+		let stats = await runWebpack(true, showStats);
 		showStats(stats);
 	}
 });
