+++ src/lib/webpack/npm-install-loader.js
